requirements.txt
requests
beautifulsoup4
sqlalchemy[asyncio]
asyncpg
httpx
python-dotenv